Add interface for frame writer
