+++ webm/framewriter.go
@@ -15,11 +15,17 @@
 package webm
 
 import (
+	"io"
+	"os"
 	"sync"
+
+	"github.com/at-wat/ebml-go"
 )
 
-// FrameWriter is a stream frame writer.
-// It's simillar to io.Writer, but having additional arguments keyframe flag and timestamp.
+// FrameWriter is an implementation of WriteCloser.
+//
+// Deprecated: This is exposed to keep compatibility with the old version.
+// Use WriteCloser interface instead.
 type FrameWriter struct {
 	trackNumber uint64
 	f           chan *frame
@@ -55,4 +61,173 @@ func (w *FrameWriter) Close() error {
 	w.fin <- struct{}{}
 
 	return nil
+}
+
+// NewFrameWriter creates WriteCloser for each track specified as tracks argument.
+// Resultant WebM is written to given io.WriteCloser.
+// io.WriteCloser will be closed automatically; don't close it by yourself.
+func NewFrameWriter(w0 io.WriteCloser, tracks []TrackEntry, opts ...FrameWriterOption) ([]WriteCloser, error) {
+	options := &FrameWriterOptions{
+		ebmlHeader:  DefaultEBMLHeader,
+		segmentInfo: DefaultSegmentInfo,
+	}
+	for _, o := range opts {
+		if err := o(options); err != nil {
+			return nil, err
+		}
+	}
+
+	w := &writerWithSizeCount{w: w0}
+
+	type FlexSegment struct {
+		SeekHead interface{} `ebml:"SeekHead,omitempty"`
+		Info     interface{} `ebml:"Info"`
+		Tracks   Tracks      `ebml:"Tracks"`
+		Cluster  []Cluster   `ebml:"Cluster"`
+	}
+
+	header := struct {
+		Header  interface{} `ebml:"EBML"`
+		Segment FlexSegment `ebml:"Segment,size=unknown"`
+	}{
+		Header: options.ebmlHeader,
+		Segment: FlexSegment{
+			SeekHead: options.seekHead,
+			Info:     options.segmentInfo,
+			Tracks: Tracks{
+				TrackEntry: tracks,
+			},
+		},
+	}
+	if err := ebml.Marshal(&header, w, options.marshalOpts...); err != nil {
+		return nil, err
+	}
+
+	w.Clear()
+
+	ch := make(chan *frame)
+	fin := make(chan struct{}, len(tracks)-1)
+	wg := sync.WaitGroup{}
+	var ws []WriteCloser
+
+	for _, t := range tracks {
+		wg.Add(1)
+		ws = append(ws, &FrameWriter{
+			trackNumber: t.TrackNumber,
+			f:           ch,
+			wg:          &wg,
+			fin:         fin,
+		})
+	}
+
+	closed := make(chan struct{})
+	go func() {
+		wg.Wait()
+		close(closed)
+	}()
+
+	go func() {
+		const invalidTimestamp = int64(0x7FFFFFFFFFFFFFFF)
+		tc0 := invalidTimestamp
+		tc1 := invalidTimestamp
+		lastTc := int64(0)
+
+		defer func() {
+			// Finalize WebM
+			if tc0 == invalidTimestamp {
+				// No data written
+				tc0 = 0
+			}
+			cluster := struct {
+				Cluster Cluster `ebml:"Cluster,size=unknown"`
+			}{
+				Cluster: Cluster{
+					Timecode: uint64(lastTc - tc0),
+					PrevSize: uint64(w.Size()),
+				},
+			}
+			if err := ebml.Marshal(&cluster, w, options.marshalOpts...); err != nil {
+				// TODO: output error
+				panic(err)
+			}
+			w.Close()
+			<-fin // read one data to release blocked Close()
+		}()
+
+	L_WRITE:
+		for {
+			select {
+			case <-closed:
+				break L_WRITE
+			case f := <-ch:
+				if tc0 == invalidTimestamp {
+					tc0 = f.timestamp
+				}
+				lastTc = f.timestamp
+				tc := f.timestamp - tc1
+				if tc >= 0x7FFF || tc1 == invalidTimestamp {
+					// Create new Cluster
+					tc1 = f.timestamp
+					tc = 0
+
+					cluster := struct {
+						Cluster Cluster `ebml:"Cluster,size=unknown"`
+					}{
+						Cluster: Cluster{
+							Timecode: uint64(tc1 - tc0),
+							PrevSize: uint64(w.Size()),
+						},
+					}
+					w.Clear()
+					if err := ebml.Marshal(&cluster, w, options.marshalOpts...); err != nil {
+						// TODO: output error
+						panic(err)
+					}
+				}
+				if tc <= -0x7FFF {
+					// Ignore too old frame
+					// TODO: output error
+					continue
+				}
+
+				b := struct {
+					Block ebml.Block `ebml:"SimpleBlock"`
+				}{
+					ebml.Block{
+						TrackNumber: f.trackNumber,
+						Timecode:    int16(tc),
+						Keyframe:    f.keyframe,
+						Data:        [][]byte{f.b},
+					},
+				}
+				// Write SimpleBlock to the file
+				if err := ebml.Marshal(&b, w, options.marshalOpts...); err != nil {
+					// TODO: output error
+					panic(err)
+				}
+			}
+		}
+	}()
+
+	return ws, nil
+}
+
+// NewSimpleWriter creates WriteCloser for each track specified as tracks argument.
+// Resultant WebM is written to given io.WriteCloser.
+// io.WriteCloser will be closed automatically; don't close it by yourself.
+//
+// Deprecated: This is exposed to keep compatibility with the old version.
+// Use NewFrameWriter instead.
+func NewSimpleWriter(w0 io.WriteCloser, tracks []TrackEntry, opts ...FrameWriterOption) ([]*FrameWriter, error) {
+	os.Stderr.WriteString(
+		"Deprecated: You are using deprecated webm.NewSimpleWriter and *webm.FrameWriter.\n" +
+			"            Use webm.NewFrameWriter and webm.WriteCloser instead.\n" +
+			"            See https://godoc.org/github.com/at-wat/ebml-go to find out the latest API.\n",
+	)
+	ws, err := NewFrameWriter(w0, tracks, opts...)
+	var ws2 []*FrameWriter
+	for _, w := range ws {
+		ws2 = append(ws2, w.(*FrameWriter))
+	}
+	return ws2, err
 }
